requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-spinner==0.1.10
colorama==0.4.6
croniter==1.4.1
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.0
fastapi_amis_admin==0.7.2
fastapi_scheduler==0.0.15
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
mailersend==0.5.6
Mako==1.3.3
MarkupSafe==2.1.5
orjson==3.9.15
packaging==24.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlalchemy_database==0.1.2
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
